Implemented functionality for reviewer to view all applications, view all applications of a specific type along with exceptions + error handling for the most likely errors, rewrote application to hopefully be correct from a database perspective + added setters + changed version number to long, added application repository + updated database loader to add test application + fixed various comments

+++ SpringProgram/demo/src/main/java/com/example/demo/domain/dto/ApplicationDTO.java
@@ -37,7 +37,7 @@ public interface ApplicationDTO {
      * 
      * @return the version number for this application
      */
-    public Integer getVersionNumber();
+    public long getVersionNumber();
 
     /** Interface function for retriving the application status for this application
      *
